Ducnv

+++ ManagerShop/src/com/raven/form/FormCustomer.java
@@ -9,7 +9,10 @@ import com.fpt.DAO.CustomerDAO;
 import com.fpt.Validate.Validate;
 import com.fpt.Validate.labelValidate;
 import com.fpt.entity.Customer;
+import com.fpt.entity.User;
+import com.fpt.utils.MsgBox;
 import com.raven.dialog.Message;
+import java.awt.event.ActionListener;
 import java.util.List;
 import javax.swing.JOptionPane;
 import javax.swing.table.DefaultTableModel;
@@ -43,18 +46,20 @@ public class FormCustomer extends javax.swing.JPanel {
             model.addRow(row);
         }
     }
+
     public void fillTableWhenFind() {
         DefaultTableModel model = (DefaultTableModel) tableCustomer.getModel();
         model.setRowCount(0);
-        String keyword = txtTimkiem.getText();
-        List<Customer> list = cDao.selectByKeyWord(keyword);
-        if(list.size() == 0){
-            lblTimKiem.setText("Không có dữ liệu khách hàng nào được tìm thấy");
+        String keyString = txtTimkiem.getText();
+        List<Customer> list = cDao.selectByKeyWord(keyString);
+        if (list.isEmpty()) {
+            lblTimKiem.setText("Không có khách hàng " + keyString);
             return;
         }
         for (Customer c : list) {
-            Object[] row = {c.getId(), c.getName(), c.getAddress(), c.getPhoneNumber(), c.getGender() ? "Nam" : "Nu"};
-            model.addRow(row);
+            model.addRow(new Object[]{
+                c.getId(), c.getName(), c.getAddress(), c.getPhoneNumber(), c.getGender() ? "Nam" : "Nu"
+            });
         }
         lblTimKiem.setText("");
     }
@@ -125,13 +130,15 @@ public class FormCustomer extends javax.swing.JPanel {
     public void delete() {
         int row = tableCustomer.getSelectedRow();
         int ma = (int) tableCustomer.getValueAt(row, 0);
-        try {
-            cDao.delete(ma);
-            fillTable();
-            clearForm();
-            JOptionPane.showMessageDialog(this, "Xoa thanh cong");
-        } catch (Exception e) {
-            e.printStackTrace();
+        if (MsgBox.confirm(this, "Bạn có muốn xóa không?")) {
+            try {
+                cDao.delete(ma);
+                fillTable();
+                clearForm();
+                JOptionPane.showMessageDialog(this, "Xoa thanh cong");
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
     }
 
@@ -144,7 +151,10 @@ public class FormCustomer extends javax.swing.JPanel {
         Customer c = cDao.selectById(ma);
         setForm(c);
     }
-
+    
+    public void addEvenFillTable(ActionListener evt){
+        fillTable();
+    }
     /**
      * This method is called from within the constructor to initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is always
@@ -161,9 +171,9 @@ public class FormCustomer extends javax.swing.JPanel {
         btnTim = new com.raven.suportSwing.MyButton();
         lblTimKiem = new javax.swing.JLabel();
         jPanel2 = new javax.swing.JPanel();
+        scrollBar1 = new com.raven.suportSwing.ScrollBar();
         jScrollPane1 = new javax.swing.JScrollPane();
         tableCustomer = new com.raven.suportSwing.TableColumn();
-        scrollBar1 = new com.raven.suportSwing.ScrollBar();
         jPanel3 = new javax.swing.JPanel();
         txtName = new com.raven.suportSwing.TextField();
         btnClear = new com.raven.suportSwing.MyButton();
@@ -187,6 +197,21 @@ public class FormCustomer extends javax.swing.JPanel {
         jLabel2.setText("Khách hàng");
 
         txtTimkiem.setLabelText("Tìm theo tên khách hàng");
+        txtTimkiem.addFocusListener(new java.awt.event.FocusAdapter() {
+            public void focusGained(java.awt.event.FocusEvent evt) {
+                txtTimkiemFocusGained(evt);
+            }
+        });
+        txtTimkiem.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                txtTimkiemActionPerformed(evt);
+            }
+        });
+        txtTimkiem.addKeyListener(new java.awt.event.KeyAdapter() {
+            public void keyReleased(java.awt.event.KeyEvent evt) {
+                txtTimkiemKeyReleased(evt);
+            }
+        });
 
         btnTim.setText("Tìm");
         btnTim.setRadius(20);
@@ -207,7 +232,7 @@ public class FormCustomer extends javax.swing.JPanel {
                 .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 129, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(txtTimkiem, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(txtTimkiem, javax.swing.GroupLayout.DEFAULT_SIZE, 299, Short.MAX_VALUE)
                     .addComponent(lblTimKiem, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
                 .addGap(31, 31, 31)
                 .addComponent(btnTim, javax.swing.GroupLayout.PREFERRED_SIZE, 64, javax.swing.GroupLayout.PREFERRED_SIZE)
@@ -231,19 +256,22 @@ public class FormCustomer extends javax.swing.JPanel {
 
         jPanel2.setBackground(new java.awt.Color(255, 255, 255));
 
-        jScrollPane1.setVerticalScrollBar(scrollBar1);
-
         tableCustomer.setModel(new javax.swing.table.DefaultTableModel(
             new Object [][] {
-                {null, null, null, null, null},
-                {null, null, null, null, null},
-                {null, null, null, null, null},
-                {null, null, null, null, null}
+
             },
             new String [] {
-                "ID", "Tên KH", "Địa chỉ", "SĐT", "Giới tính"
+                "ID ", "Tên KH", "Địa Chỉ", "SĐT", "Giới tính"
             }
-        ));
+        ) {
+            boolean[] canEdit = new boolean [] {
+                false, false, false, false, false
+            };
+
+            public boolean isCellEditable(int rowIndex, int columnIndex) {
+                return canEdit [columnIndex];
+            }
+        });
         tableCustomer.addMouseListener(new java.awt.event.MouseAdapter() {
             public void mouseClicked(java.awt.event.MouseEvent evt) {
                 tableCustomerMouseClicked(evt);
@@ -256,22 +284,25 @@ public class FormCustomer extends javax.swing.JPanel {
         jPanel2Layout.setHorizontalGroup(
             jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel2Layout.createSequentialGroup()
-                .addContainerGap(11, Short.MAX_VALUE)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 673, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addComponent(scrollBar1, javax.swing.GroupLayout.PREFERRED_SIZE, 5, javax.swing.GroupLayout.PREFERRED_SIZE))
+            .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(jPanel2Layout.createSequentialGroup()
+                    .addContainerGap()
+                    .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 697, Short.MAX_VALUE)
+                    .addContainerGap()))
         );
         jPanel2Layout.setVerticalGroup(
             jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel2Layout.createSequentialGroup()
-                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel2Layout.createSequentialGroup()
-                        .addGap(42, 42, 42)
-                        .addComponent(scrollBar1, javax.swing.GroupLayout.PREFERRED_SIZE, 148, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(jPanel2Layout.createSequentialGroup()
-                        .addGap(25, 25, 25)
-                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 298, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addGap(42, 42, 42)
+                .addComponent(scrollBar1, javax.swing.GroupLayout.PREFERRED_SIZE, 148, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+            .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(jPanel2Layout.createSequentialGroup()
+                    .addContainerGap()
+                    .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 486, Short.MAX_VALUE)
+                    .addContainerGap()))
         );
 
         jPanel3.setBackground(new java.awt.Color(255, 255, 255));
@@ -419,7 +450,7 @@ public class FormCustomer extends javax.swing.JPanel {
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(jPanel2, javax.swing.GroupLayout.DEFAULT_SIZE, 684, Short.MAX_VALUE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jPanel3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
         );
@@ -429,7 +460,7 @@ public class FormCustomer extends javax.swing.JPanel {
                 .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(0, 0, 0)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jPanel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jPanel2, javax.swing.GroupLayout.DEFAULT_SIZE, 349, Short.MAX_VALUE)
                     .addComponent(jPanel3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))
         );
     }// </editor-fold>//GEN-END:initComponents
@@ -473,11 +504,6 @@ public class FormCustomer extends javax.swing.JPanel {
 
     }//GEN-LAST:event_btnThemActionPerformed
 
-    private void tableCustomerMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_tableCustomerMouseClicked
-        // TODO add your handling code here:
-        edit();
-    }//GEN-LAST:event_tableCustomerMouseClicked
-
     private void btnXoaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnXoaActionPerformed
         // TODO add your handling code here:
         delete();
@@ -499,6 +525,26 @@ public class FormCustomer extends javax.swing.JPanel {
         // TODO add your handling code here:
         fillTableWhenFind();
     }//GEN-LAST:event_btnTimActionPerformed
+
+    private void tableCustomerMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_tableCustomerMouseClicked
+        // TODO add your handling code here:
+        edit();
+    }//GEN-LAST:event_tableCustomerMouseClicked
+
+    private void txtTimkiemActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtTimkiemActionPerformed
+        // TODO add your handling code here:
+        fillTableWhenFind();
+    }//GEN-LAST:event_txtTimkiemActionPerformed
+
+    private void txtTimkiemFocusGained(java.awt.event.FocusEvent evt) {//GEN-FIRST:event_txtTimkiemFocusGained
+        // TODO add your handling code here:
+        fillTableWhenFind();
+    }//GEN-LAST:event_txtTimkiemFocusGained
+
+    private void txtTimkiemKeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_txtTimkiemKeyReleased
+        // TODO add your handling code here:
+        fillTableWhenFind();
+    }//GEN-LAST:event_txtTimkiemKeyReleased
 
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
